Sort out Tree module ...
This might be temporay though based on WMA docs
